gerenciar logs

+++ others/deletar.js
@@ -1,11 +1,4 @@
-/*const { Client, GatewayIntentBits } = require('discord.js');
-const client = new Client({ 
-    intents: [
-        GatewayIntentBits.Guilds,
-        GatewayIntentBits.GuildMessages,
-        GatewayIntentBits.MessageContent
-    ],
-});
+const client = require('../index');
 const chalk = require('chalk');
 
 require('dotenv').config({ path: './secure/.env' });
@@ -15,29 +8,36 @@ function apagaTudo(){
         const canalDeploy = client.channels.cache.get(ID_LOGS_DEPLOY);
         const apagaDeploy = canalDeploy.messages.fetch()
         .then((mensagens) => {
-            canalDeploy.bulkDelete(mensagens);
-            console.log(chalk.yellow(`Foram apagadas ${mensagens.size} mensagens! No canal ${canalDeploy.name}.`));
-            canalDeploy.send("Mensagens Apagadas!");
+            if(mensagens.size <= 12){
+                canalDeploy.bulkDelete(mensagens);
+                console.log(chalk.yellow(`Foram apagadas ${mensagens.size} mensagens! No canal ${canalDeploy.name}.`));
+                canalDeploy.send(`Foram apagadas ${mensagens.size} mensagens! No canal ${canalDeploy.name}.`);
+            }
         })
         .catch(console.error);
-        return apagaDeploy;
 
         const canalInfo = client.channels.cache.get(ID_LOGS_INFO);
         const apagaInfo = canalInfo.messages.fetch()
         .then((mensagens) => {
-            return canalInfo.bulkDelete(mensagens);
-            //console.log(chalk.yellow(`Foram apagadas ${mensagens.size} mensagens! No canal ${mensagens.channel.name}.`));
+            if(mensagens.size <= 30){
+                canalCommands.bulkDelete(mensagens);
+                console.log(chalk.yellow(`Foram apagadas ${mensagens.size} mensagens! No canal ${canalInfo.name}.`));
+                canalInfo.send(`Foram apagadas ${mensagens.size} mensagens! No canal ${canalInfo.name}.`);
+            }
         })
         .catch(console.error);
 
         const canalCommands = client.channels.cache.get(ID_LOGS_COMMANDS);
         const apagaCommands = canalCommands.messages.fetch()
         .then((mensagens) => {
-            return canalCommands.bulkDelete(mensagens);
-            //console.log(chalk.yellow(`Foram apagadas ${mensagens.size} mensagens! No canal ${mensagens.channel.name}.`));
+            if(mensagens.size <= 30){
+                canalCommands.bulkDelete(mensagens);
+                console.log(chalk.yellow(`Foram apagadas ${mensagens.size} mensagens! No canal ${canalCommands.name}.`));
+                canalCommands.send(`Foram apagadas ${mensagens.size} mensagens! No canal ${canalCommands.name}.`);
+            }
         })
         .catch(console.error);
         return apagaCommands, apagaDeploy, apagaInfo;
     }
 module.exports = apagaTudo;
-*/
+
